+++ metadata.rb
@@ -1,7 +1,8 @@
 name             'netapp'
-maintainer       'YOUR_COMPANY_NAME'
-maintainer_email 'YOUR_EMAIL'
+maintainer       'Blue Medora EngOps Department'
+maintainer_email 'engops@bluemedora.com'
 license          'All rights reserved'
-description      'Installs/Configures netapp'
+description      'Installs/Configures NetApp Software'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'
+depends          'yum-mysql-community'