+++ recipes/api-slm.rb
@@ -1,0 +1,6 @@
+##
+# This attributes file contains all instructions necessary to install API Services or SLM
+# Authors:
+#   Dylan Myers
+# Copyright 2017 Blue Medora
+##