+++ recipes/ocum.rb
@@ -1,0 +1,6 @@
+##
+# This attributes file contains all instructions necessary to install On Command Unified Manager
+# Authors:
+#   Dylan Myers
+# Copyright 2017 Blue Medora
+##